% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{body_insert}
\alias{body_insert}
\title{Insert Into a Function Body}
\usage{
body_insert(fn_body, insertion, before = NULL)
}
\arguments{
\item{fn_body}{The body of a function (as found via body(fun)).}

\item{insertion}{A quoted expression to add at the beginning of the function.}
}
\value{
A function body with the insertion. Note: If before is specified and
  is not found anywhere in fn_body, fn_body is returned unaltered.
}
\description{
Insert quoted insertions at the start of a function body (after the opening
of the function).
}
\examples{
fun <- function(x) {
  x + 1
}
body_insert(body(fun), quote(x + 2))
}
