% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidy-eval.R
\name{:=}
\alias{:=}
\title{Digested is}
\arguments{
\item{lhs}{An expression that evaluates to a character or a symbol (used as a
function parameter).}

\item{rhs}{The thing to assign to that parameter.}
}
\description{
Tidy evaluation uses a special parameter assignment operator, \code{:=}. See
\code{\link[rlang]{quasiquotation}} for more information.
}
\keyword{internal}
