% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory-package.R
\docType{package}
\name{factory-package}
\alias{factory}
\alias{factory-package}
\title{factory: Build Function Factories}
\description{
Function factories are functions that make functions. They can be 
    confusing to construct. Straightforward techniques can produce functions 
    that are fragile or hard to understand. While more robust techniques exist 
    to construct function factories, those techniques can be confusing. This 
    package is designed to make it easier to construct function factories.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/factory}
  \item Report bugs at \url{https://github.com/jonthegeek/factory/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com}

}
\keyword{internal}
