% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory.R
\name{factory}
\alias{factory}
\title{Easily Build Function Factories}
\usage{
factory(fun, ...,
  .error_message = "One or more arguments to this factory are missing.")
}
\arguments{
\item{fun}{An anonymous function to turn into a factory.}

\item{...}{Arguments for the factory function. If the argument should not
have a default, use the somewhat strange construction \code{a =}. This also
applies for \code{...}, so \code{... =}. Things on the RHS will be
evaluated before building your factory unless explicitly quoted with
\code{quote}. See examples.}

\item{.error_message}{An optional error message for the factory to use when a
required argument is not provided.}
}
\value{
A function factory.
}
\description{
Easily Build Function Factories
}
\examples{
y <- 3
power <- factory(\\{
  function(x) \\{
    x ^ exp
  \\}
\\},
exp =,
.error_message = "Please provide an exponent for your function as exp.")
square <- power()
square(2)
square(y)
y <- 7
square(y)
}
