% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factory-package.R
\docType{package}
\name{factory-package}
\alias{factory}
\alias{factory-package}
\title{factory: Easily Build Function Factories}
\description{
Easily build function factories that act as "normal" functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/factory}
  \item Report bugs at \url{https://github.com/jonthegeek/factory/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jon.harmon@macmillan.com}

}
\keyword{internal}
